feat: Flesh out settings to config Rename project Fix all warnings
