+++ config.py
@@ -1,0 +1,4 @@
+import logging
+
+loglevel = logging.INFO
+user_agent = "EntityValidator (https://github.com/dpriskorn/entityshape)"
