--- entityshape/models/property_value.py
@@ -1,11 +1,0 @@
-from typing import Optional
-
-from pydantic import BaseModel
-
-from entityshape.enums import Necessity, PropertyResponse
-
-
-class PropertyValue(BaseModel):
-    name: str = ""
-    necessity: Necessity
-    response: Optional[PropertyResponse] = None
