+++ entityvalidator/exceptions.py
