+++ entityvalidator/models/entity.py
@@ -1,25 +1,24 @@
 import re
 from re import Pattern
-from typing import Any, Dict
+from typing import Any
 
 import requests
 from pydantic import BaseModel
 
-from entityshape.exceptions import EntityIdError, LangError
-from entityshape.models.compareshape import CompareShape
-from entityshape.models.result import Result
-from entityshape.models.shape import Shape
+from entityvalidator.exceptions import EntityIdError, LangError
+from entityvalidator.models.compareshape import CompareShape
+from entityvalidator.models.result import Result
+from entityvalidator.models.shape import Shape
 
 
 class Entity(BaseModel):
     entity_id: str
     entity_id_regex: Pattern = re.compile(r"[QL]\d+")
-    entity_data: Dict[str, Any]
-    entity_schema_data: Dict[str, Any]
+    entity_data: dict[str, Any]
+    entity_schema_data: dict[str, Any]
     eid: str  # entityshape
-    lang: str = "en"  # language defaults to English
     result: Result = Result()
-    compare_shape_result: Dict[str, Any] = {}
+    compare_shape_result: dict[str, Any] = {}
     wikibase_url: str = "http://www.wikidata.org"
     mediawiki_api_url: str = "https://www.wikidata.org/w/api.php"
     user_agent: str = "entityshape (https://github.com/dpriskorn/entityshape)"
@@ -34,7 +33,7 @@ class Entity(BaseModel):
 
     def __validate__(self):
         shape: Shape = Shape(
-            entity_schema_id=self.eid,
+            # entity_schema_id=self.eid,
             language=self.lang,
             entity_schema_data=self.entity_schema_data,
         )
@@ -42,7 +41,6 @@ class Entity(BaseModel):
             shape=shape.get_schema_shape(),
             entity=self.entity_id,
             entity_data=self.entity_data,
-            language=self.lang,
             wikibase_url=self.wikibase_url,
             mediawiki_api_url=self.mediawiki_api_url,
         )
@@ -70,3 +68,8 @@ class Entity(BaseModel):
         url = f"{self.wikibase_url}/wiki/Special:EntityData/{self.entity_id}.json"
         response = requests.get(url)
         self.entity_data = response.json()
+
+    def to_dict(self) -> dict[Any, Any]:
+        json_ = self.result.to_dict()
+        json_["entity"] = self.entity_id
+        return json_
