+++ entityvalidator/models/property_value.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+from entityvalidator.enums import Necessity, PropertyResponse
+
+
+class PropertyValue(BaseModel):
+    name: str = ""
+    necessity: Necessity
+    response: PropertyResponse | None = None
