+++ entityvalidator/models/result_property.py
